fixes, modularize
